1.elasticsearch client增加bulkprocessor功能
2.增加bulkprocessor功能demo TestBulkProcessor
https://github.com/bbossgroups/elasticsearch-example/blob/master/src/test/java/org/bboss/elasticsearchtest/bulkprocessor/TestBulkProcessor.java

3.调整版本号为5.9.7

+++ bboss-elasticsearch-rest-kafka/src/main/java/org/frameworkset/tran/kafka/KafkaContext.java
@@ -35,7 +35,7 @@ public interface KafkaContext {
 	public int getConsumerThreads() ;
 
 
-	public int getPollTimeOut() ;
+	public long getPollTimeOut() ;
 	public String getValueCodec();
 	public String getKeyCodec();
