+++ bboss-elasticsearch-rest-kafka/src/main/java/org/frameworkset/tran/kafka/KafkaImportContext.java
@@ -70,7 +70,7 @@ public abstract class KafkaImportContext extends BaseImportContext implements Ka
 
 
 
-	public int getPollTimeOut() {
+	public long getPollTimeOut() {
 		return kafkaImportConfig.getPollTimeOut();
 	}
